arka_plan_silici_flask/requirements.txt
flask ==1.1.1
pillow ==7.0.0
torch ==1.6.0
torchvision==0.7.0
gunicorn==20.0.4